+++ src/main/java/com/milaboratory/mist/pattern/PlusPattern.java
@@ -52,7 +52,7 @@ public class PlusPattern extends MultiplePatternsOperator {
 
         @Override
         protected void performSearch(boolean quickSearch) {
-            int bestScore = 0;
+            float bestScore = 0;
             int numOperands = operandPatterns.length;
 
             // initialize arrays and get matches for all operands
@@ -105,7 +105,7 @@ public class PlusPattern extends MultiplePatternsOperator {
                     }
                     Match currentMatch = combineMatches(input, targetId, currentMatches);
                     if (!quickBestMatchFound) {
-                        int currentSum = sumMatchesScore(currentMatches);
+                        float currentSum = sumMatchesScore(currentMatches);
                         if (currentSum > bestScore) {
                             bestMatch = currentMatch;
                             bestScore = currentSum;
